Refactored code into modules.py and predict.py and made comments more readable
