--- recognition/maskRNN/module.py
@@ -1,17 +1,0 @@
-import torch
-import torchvision
-from torchvision.models.detection import MaskRCNN
-from torchvision.models.detection.faster_rcnn import FastRCNNPredictor
-from torchvision.models.detection.mask_rcnn import MaskRCNNPredictor
-from torchvision import transforms
-
-def loadModel():
-    model = torchvision.models.detection.maskrcnn_resnet50_fpn(weights="DEFAULT", min_size=1000,max_size=4288) #Why
-
-    num_classes = 2  
-    in_features = model.roi_heads.box_predictor.cls_score.in_features
-    model.roi_heads.box_predictor = FastRCNNPredictor(in_features, num_classes)
-    in_features_mask = model.roi_heads.mask_predictor.conv5_mask.in_channels
-    hidden_layer = 256
-    model.roi_heads.mask_predictor = MaskRCNNPredictor(in_features_mask, hidden_layer, num_classes)
-    return model
